# ensure we error instead of silently misconstructing rows when we encounter pages of repeated columns that do not start at the beginnig of a row