--- a/.gitignore
+++ b/.gitignore
@@ -4,6 +4,7 @@
 *.dll
 *.so
 *.dylib
+*.py
 
 # Test binary, built with `go test -c`
 *.test