--- a/file_test.go
+++ b/file_test.go
@@ -5,6 +5,7 @@ import (
 	"io"
 	"os"
 	"path/filepath"
+	"slices"
 	"strings"
 	"testing"
 
@@ -12,12 +13,15 @@ import (
 )
 
 var testdataFiles []string
+var testdataFilesWithMalformedRepetitionLevels = []string{
+	"testdata/issue206.parquet",
+	"testdata/file.parquet",
+	"testdata/small.parquet",
+	"testdata/trace.snappy.parquet",
+}
 
 func init() {
-	entries, _ := os.ReadDir("testdata")
-	for _, e := range entries {
-		testdataFiles = append(testdataFiles, filepath.Join("testdata", e.Name()))
-	}
+	testdataFiles, _ = filepath.Glob("testdata/*.parquet")
 }
 
 func TestOpenFile(t *testing.T) {
@@ -49,7 +53,7 @@ func TestOpenFile(t *testing.T) {
 			parquet.PrintSchema(b, root.Name(), root)
 			t.Log(b)
 
-			printColumns(t, p.Root(), "")
+			printColumns(t, p.Root(), "", path)
 		})
 	}
 }
@@ -121,7 +125,7 @@ func TestOpenFileWithoutPageIndex(t *testing.T) {
 	}
 }
 
-func printColumns(t *testing.T, col *parquet.Column, indent string) {
+func printColumns(t *testing.T, col *parquet.Column, indent, filePath string) {
 	if t.Failed() {
 		return
 	}
@@ -163,11 +167,18 @@ func printColumns(t *testing.T, col *parquet.Column, indent string) {
 			}
 			numValues += int64(n)
 			if err != nil {
-				if err != io.EOF {
-					t.Error(err)
-					return
+				if err == io.EOF {
+					break
 				}
-				break
+				if errors.Is(err, parquet.ErrMalformedRepetitionLevel) {
+					if slices.Contains(testdataFilesWithMalformedRepetitionLevels, filePath) {
+						numValues = p.NumValues()
+						nullCount = p.NumNulls()
+						break
+					}
+				}
+				t.Error(err)
+				return
 			}
 		}
 
@@ -185,7 +196,7 @@ func printColumns(t *testing.T, col *parquet.Column, indent string) {
 	}
 
 	for _, child := range col.Columns() {
-		printColumns(t, child, indent)
+		printColumns(t, child, indent, filePath)
 	}
 }