--- a/reader_test.go
+++ b/reader_test.go
@@ -863,3 +863,79 @@ func TestLookup(t *testing.T) {
 		}
 	})
 }
+
+// struct for parquet-go/parquet-go
+type OrderbookDepth struct {
+	Timestamp  int64   `parquet:"timestamp"       json:"timestamp"`
+	Instrument string  `parquet:"instrument_name" json:"instrument_name"`
+	ChangeID   int64   `parquet:"change_id"       json:"change_id"`
+	Bids       []Level `parquet:"bids,list"       json:"bids"`
+	Asks       []Level `parquet:"asks,list"       json:"asks"`
+}
+
+type Level struct {
+	Price  float64 `parquet:"price"  json:"price"`
+	Amount float64 `parquet:"amount" json:"amount"`
+}
+
+func TestIssue206(t *testing.T) {
+	f, err := os.Open("testdata/issue206.parquet")
+	if err != nil {
+		t.Fatalf("Failed to open file: %v", err)
+	}
+	defer f.Close()
+	s, err := f.Stat()
+	if err != nil {
+		t.Fatalf("Failed to get file stats: %v", err)
+	}
+
+	if _, err := parquet.Read[OrderbookDepth](f, s.Size()); !errors.Is(err, parquet.ErrMalformedRepetitionLevel) {
+		t.Fatal(err)
+	}
+
+	// TODO: add these tests back if we implement a fix for issue #206
+	//
+	// rows, err := parquet.Read[OrderbookDepth](f, s.Size())
+	// if err != nil {
+	// 	t.Fatalf("Failed to read file: %v", err)
+	// }
+	//
+	// b, err := os.ReadFile("testdata/issue206.json")
+	// if err != nil {
+	// 	t.Fatalf("Failed to open file: %v", err)
+	// }
+
+	// var want []OrderbookDepth
+	// if err := json.Unmarshal(b, &want); err != nil {
+	// 	t.Fatalf("Failed to unmarshal JSON: %v", err)
+	// }
+
+	// if len(rows) != len(want) {
+	// 	t.Errorf("len(rows) != len(want): %d != %d", len(rows), len(want))
+	// }
+
+	// for i := range min(len(rows), len(want)) {
+	// 	g := rows[i]
+	// 	w := want[i]
+	// 	if g.Timestamp != w.Timestamp {
+	// 		t.Errorf("rows[%d].Timestamp != want[%d].Timestamp: %d != %d", i, i, g.Timestamp, w.Timestamp)
+	// 		break
+	// 	}
+	// 	if g.Instrument != w.Instrument {
+	// 		t.Errorf("rows[%d].Instrument != want[%d].Instrument: %q != %q", i, i, g.Instrument, w.Instrument)
+	// 		break
+	// 	}
+	// 	if g.ChangeID != w.ChangeID {
+	// 		t.Errorf("rows[%d].ChangeID != want[%d].ChangeID: %d != %d", i, i, g.ChangeID, w.ChangeID)
+	// 		break
+	// 	}
+	// 	if !slices.Equal(g.Bids, w.Bids) {
+	// 		t.Errorf("rows[%d].Bids != want[%d].Bids:\ngot:  %+v\nwant: %+v", i, i, g.Bids, w.Bids)
+	// 		break
+	// 	}
+	// 	if !slices.Equal(g.Asks, w.Asks) {
+	// 		t.Errorf("rows[%d].Asks != want[%d].Asks:\ngot:  %+v\nwant: %+v", i, i, g.Asks, w.Asks)
+	// 		break
+	// 	}
+	// }
+}