Fixes #204 
Fixes #206 

It appears that parquet-go has issues reading legacy parquet files encoded with data page v1 when the pages do not start at the beginning of a row. This issue does not occur on data page v2 because the spec explicitly says that the pages must start with a new row.

~For now I've only added a check to ensure that we detect malformed files, which is much better than silently doing the wrong thing.~

Edit: I figured out how to make it work.