@achille-roussel 2025-01-27
This change has a minor performance impact on reading byte array columns but this seems like an acceptable cost for the fix.
```
goos: darwin
goarch: arm64
pkg: github.com/parquet-go/parquet-go
cpu: Apple M2 Pro
                                     │ /tmp/bench.old │           /tmp/bench.new           │
                                     │     sec/op     │   sec/op     vs base               │
GenericReader/byteArrayColumn/go1.18      55.35µ ± 0%   58.89µ ± 2%  +6.40% (p=0.000 n=10)

                                     │ /tmp/bench.old │           /tmp/bench.new           │
                                     │     row/s      │    row/s     vs base               │
GenericReader/byteArrayColumn/go1.18      16.43M ± 0%   16.98M ± 1%  +3.38% (p=0.000 n=10)

                                     │ /tmp/bench.old │           /tmp/bench.new            │
                                     │      B/op      │     B/op      vs base               │
GenericReader/byteArrayColumn/go1.18     4.926Ki ± 0%   5.418Ki ± 0%  +9.99% (p=0.000 n=10)

                                     │ /tmp/bench.old │           /tmp/bench.new           │
                                     │   allocs/op    │ allocs/op   vs base                │
GenericReader/byteArrayColumn/go1.18       818.0 ± 0%   900.0 ± 0%  +10.02% (p=0.000 n=10)
```
As a compensation, it looks like it's slightly faster on other column types:
```
goos: darwin
goarch: arm64
pkg: github.com/parquet-go/parquet-go
cpu: Apple M2 Pro
                                 │ /tmp/bench.int64.new │        /tmp/bench.int64.old        │
                                 │        sec/op        │   sec/op     vs base               │
GenericReader/int64Column/go1.18            38.68µ ± 2%   35.34µ ± 0%  -8.63% (p=0.000 n=10)

                                 │ /tmp/bench.int64.new │        /tmp/bench.int64.old        │
                                 │        row/s         │    row/s     vs base               │
GenericReader/int64Column/go1.18            25.85M ± 2%   25.72M ± 0%  -0.50% (p=0.001 n=10)

                                 │ /tmp/bench.int64.new │     /tmp/bench.int64.old      │
                                 │         B/op         │    B/op      vs base          │
GenericReader/int64Column/go1.18             3.000 ± 0%   3.000 ± 33%  ~ (p=0.474 n=10)

                                 │ /tmp/bench.int64.new │      /tmp/bench.int64.old      │
                                 │      allocs/op       │ allocs/op   vs base            │
GenericReader/int64Column/go1.18             0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
```