@joe-elliott 2025-01-29
This is currently throwing errors in a lot of tempo tests including [this one](https://github.com/grafana/tempo/blob/8e740797b573e50aa6a77a7ac06aa6cb96f0e917/tempodb/encoding/vparquet4/block_findtracebyid_test.go#L22)

```
--- FAIL: TestBackendBlockFindTraceByID (0.06s)
    block_findtracebyid_test.go:101:
        	Error Trace:	/Users/joe/dev/joe-elliott/tempo/tempodb/encoding/vparquet4/block_findtracebyid_test.go:101
        	Error:      	Received unexpected error:
        	            	error reading row from backend: EOF
        	Test:       	TestBackendBlockFindTraceByID
```

I will try to isolate it better than this test.